Make the editor ise persistAutosaveToAPI
