+++ packages/core-data/src/utils/get-record-id.js
@@ -1,0 +1,9 @@
+/**
+ * Internal dependencies
+ */
+import { DEFAULT_ENTITY_KEY } from '../entities';
+
+export default function getRecordId( entity, record ) {
+	const entityIdKey = entity.key || DEFAULT_ENTITY_KEY;
+	return record[ entityIdKey ];
+}
