+++ packages/core-data/src/utils/index.js
@@ -4,5 +4,6 @@ export { default as ifMatchingAction } from './if-matching-action';
 export { default as ifNotResolved } from './if-not-resolved';
 export { default as onSubKey } from './on-sub-key';
 export { default as replaceAction } from './replace-action';
+export { default as getRecordId } from './get-record-id';
 export { default as withWeakMapCache } from './with-weak-map-cache';
 export { default as isRawAttribute } from './is-raw-attribute';
